+++ src/main/java/Command/MusicModule/VolumeCommand.java
@@ -61,19 +61,22 @@ public class VolumeCommand implements Command {
             int volume = 50;
             try {
                 volume = Integer.parseInt(args[0]);
+                
+                if(volume < 0 || volume > 100)
+                {
+                    e.getTextChannel().sendMessage(Emoji.error + " Please enter a number between 0 to 100.").queue();
+                    return;
+                }
+
+                Music.setVolume(e, volume);
+                if(volume < 50)
+                    e.getTextChannel().sendMessage(Emoji.volume_low + " Volume set to " + volume).queue();
+                else
+                    e.getTextChannel().sendMessage(Emoji.volume_high + " Volume set to " + volume).queue();
             } catch(NumberFormatException ex){
                 e.getTextChannel().sendMessage(Emoji.error + " Please enter a valid number.").queue();
                 return;
             }
-
-            if(volume < 0 || volume > 100)
-            {
-                e.getTextChannel().sendMessage(Emoji.error + " Please enter a number between 0 to 100.").queue();
-                return;
-            }
-            
-            Music.setVolume(e, volume);
-            e.getTextChannel().sendMessage(Emoji.volume_low + " Volume set to " + volume).queue();
         }
     }